more code organization tweaks, no need to use inheritance here really. Just makes it more complex.
