+++ plugins/DebuggerCore/CMakeLists.txt
@@ -24,10 +24,10 @@ if(UNIX)
 	
 	set(DebuggerCore_SRCS
 		${DebuggerCore_SRCS}
-		unix/DebuggerCoreUNIX.cpp
-		unix/DebuggerCoreUNIX.h
 		unix/Posix.cpp
 		unix/Posix.h
+		unix/Unix.cpp
+		unix/Unix.h
 	)			
 endif()
