+++ plugins/DebuggerCore/unix/Posix.h
@@ -5,6 +5,7 @@
 #include "OSTypes.h"
 #include <cstdint>
 
+namespace DebuggerCorePlugin {
 namespace Posix {
 
 void initialize();
@@ -15,6 +16,7 @@ ssize_t read(int fd, void *buf, size_t count);
 ssize_t write(int fd, const void *buf, size_t count);
 bool wait_for_sigchld(int msecs);
 
+}
 }
 
 #endif
